renamed mouse_button_map to input_state
